requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI Integration
openai==1.10.0
anthropic==0.8.1
langchain==0.1.0
langchain-openai==0.0.5

# Data Processing
pandas==2.1.4
numpy==1.25.2
openpyxl==3.1.2
xlsxwriter==3.1.9

# Power BI Integration
requests==2.31.0
aiohttp==3.9.1
msal==1.25.0

# Database and Storage
sqlalchemy==2.0.23
aiofiles==23.2.0
redis==5.0.1
psycopg2-binary==2.9.9

# Security and Validation
cryptography==41.0.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
# Utilities
python-dotenv==1.0.0
pydantic==1.10.13
python-json-logger==2.0.7

# Production
gunicorn==21.2.0
prometheus-client==0.19.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
bandit==1.7.5
safety==3.0.1
httpx==0.25.2
